Refactor credential panel
